feat: 리펙토링, section 분리
